backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.0.1
certifi==2025.8.3
click==8.2.1
ecdsa==0.19.1
fastapi==0.116.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pillow==11.3.0
prisma==0.15.0
pyasn1==0.6.1
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
